Create linuxFirewall
This allows you to connect to your EC2 instance (this ver is for linux users as it uses iptables) and configures your firewall via the terminal directly from your codes. However, I've altered this one to read and execute the terminal commands from a .txt file. 
These codes when run, opens up a firewall table allowing you to alter them from the UI itself. 

I'll be updating it in a day, so do look out for it.

.jar files required:
JSCH.jar 
jfoenix.jar